Move most patches to Patches module
We were talking about moving these a little while ago, this actually
does so.
