+++ src/Kopernicus/Injector.cs
@@ -464,397 +464,4 @@ namespace Kopernicus
             Logger.Default.Flush();
         }
     }
-
-    [HarmonyPatch(typeof(MapSO), "ConstructBilinearCoords", new Type[]
-    {
-        typeof(float), typeof(float)
-    })]
-    public static class MapSOPPatch_Float
-    {
-        private static bool Prefix(MapSO __instance, float x, float y)
-        {
-            if (ReferenceEquals(__instance, Injector.moho_height))
-            {
-                return true;
-            }
-            // X wraps around as it is longitude.
-            x = Mathf.Abs(x - Mathf.Floor(x));
-            __instance.centerX = x * __instance._width;
-            __instance.minX = Mathf.FloorToInt(__instance.centerX);
-            __instance.maxX = Mathf.CeilToInt(__instance.centerX);
-            __instance.midX = __instance.centerX - __instance.minX;
-            if (__instance.maxX == __instance._width)
-                __instance.maxX = 0;
-
-            // Y clamps as it is latitude and the poles don't wrap to each other.
-            y = Mathf.Clamp(y, 0, 0.99999f);
-            __instance.centerY = y * __instance._height;
-            __instance.minY = Mathf.FloorToInt(__instance.centerY);
-            __instance.maxY = Mathf.CeilToInt(__instance.centerY);
-            __instance.midY = __instance.centerY - __instance.minY;
-            if (__instance.maxY >= __instance._height)
-                __instance.maxY = __instance._height - 1;
-
-            return false;
-        }
-    }
-    [HarmonyPatch(typeof(MapSO), "ConstructBilinearCoords", new Type[]
-    {
-        typeof(double), typeof(double)
-    })]
-    public static class MapSOPatch_Double
-    {
-        private static bool Prefix(MapSO __instance, double x, double y)
-        {
-            if (ReferenceEquals(__instance, Injector.moho_height))
-            {
-                return true;
-            }
-            // X wraps around as it is longitude.
-            x = Math.Abs(x - Math.Floor(x));
-            __instance.centerXD = x * __instance._width;
-            __instance.minX = (int)Math.Floor(__instance.centerXD);
-            __instance.maxX = (int)Math.Ceiling(__instance.centerXD);
-            __instance.midX = (float)__instance.centerXD - __instance.minX;
-            if (__instance.maxX == __instance._width)
-                __instance.maxX = 0;
-
-            // Y clamps as it is latitude and the poles don't wrap to each other.
-            y = Math.Min(Math.Max(y, 0), 0.99999);
-            __instance.centerYD = y * __instance._height;
-            __instance.minY = (int)Math.Floor(__instance.centerYD);
-            __instance.maxY = (int)Math.Ceiling(__instance.centerYD);
-            __instance.midY = (float)__instance.centerYD - __instance.minY;
-            if (__instance.maxY >= __instance._height)
-                __instance.maxY = __instance._height - 1;
-
-            return false;
-        }
-    }
-    [HarmonyPatch(typeof(ROCManager), "ValidateCBBiomeCombos")]
-    public static class ROCManager_ValidateCBBiomeCombos
-    {
-        private static Func<ROCManager, string, CelestialBody> ValidCelestialBody;
-        private static Func<ROCManager, CelestialBody, string, bool> ValidCBBiome;
-        static bool Prefix(ROCManager __instance)
-        {
-            List<ROCDefinition> rocDefinitions = __instance.rocDefinitions;
-            ValidCelestialBody = AccessTools.MethodDelegate<Func<ROCManager, string, CelestialBody>>(AccessTools.Method(typeof(ROCManager), "ValidCelestialBody"));
-            ValidCBBiome = AccessTools.MethodDelegate<Func<ROCManager, CelestialBody, string, bool>>(AccessTools.Method(typeof(ROCManager), "ValidCBBiome"));
-            for (int num = rocDefinitions.Count - 1; num >= 0; num--)
-            {
-                for (int num2 = rocDefinitions[num].myCelestialBodies.Count - 1; num2 >= 0; num2--)
-                {
-                    CelestialBody celestialBody = ValidCelestialBody(__instance, rocDefinitions[num].myCelestialBodies[num2].name);
-                    if (celestialBody.IsNullOrDestroyed())
-                    {
-                        Debug.LogWarningFormat("[ROCManager]: Invalid CelestialBody Name {0} on ROC Definition {1}. Removed entry.", rocDefinitions[num].myCelestialBodies[num2].name, rocDefinitions[num].type);
-                        rocDefinitions[num].myCelestialBodies.RemoveAt(num2);
-                        continue; // missing in stock code
-                    }
-                    else
-                    {
-                        for (int num3 = rocDefinitions[num].myCelestialBodies[num2].biomes.Count - 1; num3 >= 0; num3--)
-                        {
-                            if (!ValidCBBiome(__instance, celestialBody, rocDefinitions[num].myCelestialBodies[num2].biomes[num3]))
-                            {
-                                Debug.LogWarningFormat("[ROCManager]: Invalid Biome Name {0} for Celestial Body {1} on ROC Definition {2}. Removed entry.", rocDefinitions[num].myCelestialBodies[num2].biomes[num3], rocDefinitions[num].myCelestialBodies[num2].name, rocDefinitions[num].type);
-                                rocDefinitions[num].myCelestialBodies[num2].biomes.RemoveAt(num3);
-                            }
-                        }
-                    }
-                    if (rocDefinitions[num].myCelestialBodies[num2].biomes.Count == 0) // ArgumentOutOfRangeException for myCelestialBodies[num2] when the previous if evaluate to true
-                    {
-                        Debug.LogWarningFormat("[ROCManager]: No Valid Biomes for Celestial Body {0} on ROC Definition {1}. Removed entry.", rocDefinitions[num].myCelestialBodies[num2].name, rocDefinitions[num].type);
-                        rocDefinitions[num].myCelestialBodies.RemoveAt(num2);
-                    }
-                }
-                if (rocDefinitions[num].myCelestialBodies.Count == 0)
-                {
-                    Debug.LogWarningFormat("[ROCManager]: No Valid Celestial Bodies on ROC Definition {0}. Removed entry.", rocDefinitions[num].type);
-                    rocDefinitions.RemoveAt(num);
-                }
-            }
-            return false;
-        }
-    }
-    [HarmonyPatch(typeof(PQSLandControl), "OnVertexBuildHeight")]
-    public static class PQSLandControl_OnVertexBuildHeight
-    {
-        static IEnumerable<CodeInstruction> Transpiler(IEnumerable<CodeInstruction> instructions)
-        {
-            FieldInfo PQSMod_sphere_field = AccessTools.Field(typeof(PQSMod), nameof(PQSMod.sphere));
-            FieldInfo PQS_sx_field = AccessTools.Field(typeof(PQS), nameof(PQS.sx));
-            MethodInfo GetLongitudeFromSX_method = AccessTools.Method(typeof(PQSLandControl_OnVertexBuildHeight), nameof(GetLongitudeFromSX));
-
-            List<CodeInstruction> code = new List<CodeInstruction>(instructions);
-
-            for (int i = 0; i < code.Count - 1; i++)
-            {
-                if (code[i].opcode == OpCodes.Ldfld && ReferenceEquals(code[i].operand, PQSMod_sphere_field)
-                                                    && code[i + 1].opcode == OpCodes.Ldfld && ReferenceEquals(code[i + 1].operand, PQS_sx_field))
-                {
-                    code[i + 1].opcode = OpCodes.Call;
-                    code[i + 1].operand = GetLongitudeFromSX_method;
-                }
-            }
-
-            return code;
-        }
-
-        /// <summary>
-        /// Transform the from the sx [-0.25, 0.75] longitude range convention where [-0.25, 0] maps to [270°, 360°]
-        /// and [0, 0.75] maps to [0°, 270°] into a linear [0,1] longitude range.
-        /// </summary>
-        public static double GetLongitudeFromSX(PQS sphere)
-        {
-            if (sphere.sx < 0.0)
-                return sphere.sx + 1.0;
-            return sphere.sx;
-        }
-    }
-
-    [HarmonyPatch(typeof(SpaceCenterCamera2), "Start")]
-    public static class SpaceCenterCamera2_Start
-    {
-        static bool Prefix(SpaceCenterCamera2 __instance)
-        {
-            CelestialBody homeBody = FlightGlobals.GetHomeBody();
-            __instance.pqsName = homeBody.pqsController.gameObject.name;
-            __instance.pqs = homeBody.pqsController;
-            __instance.t = __instance.transform;
-
-            GameEvents.onGameSceneLoadRequested.Add(__instance.OnSceneSwitch);
-
-            PQSCity ksc = homeBody.pqsController.transform.Find("KSC").GetComponent<PQSCity>();
-            __instance.altitudeInitial = (float)(ksc.planetRelativePosition.magnitude - homeBody.Radius) * -1f;
-
-            __instance.initialPosition = __instance.pqs.transform.Find(__instance.initialPositionTransformName);
-            if (__instance.initialPosition == null)
-            {
-                Debug.LogError("SpaceCenterCamera: Cannot find transform of name '" + __instance.initialPositionTransformName + "'");
-                return false;
-            }
-            __instance.t.NestToParent(__instance.initialPosition);
-            __instance.cameraTransform = new GameObject("CameraTransform").transform;
-            __instance.cameraTransform.NestToParent(__instance.transform);
-            FlightCamera.fetch.transform.NestToParent(__instance.cameraTransform);
-            FlightCamera.fetch.updateActive = false;
-            FlightCamera.fetch.gameObject.SetActive(value: true);
-            __instance.ResetCamera();
-            __instance.srfPivot = SurfaceObject.Create(__instance.initialPosition.gameObject, FlightGlobals.currentMainBody, 3, KFSMUpdateMode.FIXEDUPDATE);
-            return false;
-        }
-    }
-
-    [HarmonyPatch(typeof(GameSettings), "WriteCfg")]
-    public static class GameSettings_WriteCfg
-    {
-        static bool Prefix(GameSettings __instance)
-        {
-            PQSCache.PQSPreset pqspreset = new PQSCache.PQSPreset();
-            pqspreset.name = "Low";
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("Kerbin", 4.0, 1, 8));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("KerbinOcean", 4.0, 1, 7));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("Mun", 4.0, 1, 7));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("Minmus", 4.0, 1, 6));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("Bop", 4.0, 1, 6));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("Duna", 4.0, 1, 7));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("Eve", 4.0, 1, 7));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("EveOcean", 4.0, 1, 7));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("Gilly", 4.0, 1, 6));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("Ike", 4.0, 1, 7));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("Laythe", 4.0, 1, 7));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("LaytheOcean", 4.0, 1, 7));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("Moho", 4.0, 1, 7));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("Tylo", 4.0, 1, 7));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("Vall", 4.0, 1, 7));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("Dres", 4.0, 1, 7));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("Pol", 4.0, 1, 7));
-            pqspreset.spherePresets.Add(new PQSCache.PQSSpherePreset("Eeloo", 4.0, 1, 7));
-            PQSCache.PQSPreset pqspreset2 = new PQSCache.PQSPreset();
-            pqspreset2.name = "Default";
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("Kerbin", 6.0, 1, 9));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("KerbinOcean", 6.0, 1, 7));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("Mun", 6.0, 1, 8));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("Minmus", 6.0, 1, 6));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("Bop", 6.0, 1, 6));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("Duna", 6.0, 1, 8));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("Eve", 6.0, 1, 9));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("EveOcean", 6.0, 1, 7));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("Gilly", 6.0, 1, 6));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("Ike", 6.0, 1, 6));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("Laythe", 6.0, 1, 9));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("LaytheOcean", 6.0, 1, 7));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("Moho", 6.0, 1, 8));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("Tylo", 6.0, 1, 8));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("Vall", 6.0, 1, 8));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("Dres", 6.0, 1, 8));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("Pol", 6.0, 1, 8));
-            pqspreset2.spherePresets.Add(new PQSCache.PQSSpherePreset("Eeloo", 6.0, 1, 8));
-            PQSCache.PQSPreset pqspreset3 = new PQSCache.PQSPreset();
-            pqspreset3.name = "High";
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("Kerbin", 8.0, 1, 10));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("KerbinOcean", 8.0, 1, 7));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("Mun", 8.0, 1, 9));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("Minmus", 8.0, 1, 7));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("Bop", 8.0, 1, 6));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("Duna", 8.0, 1, 9));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("Eve", 8.0, 1, 10));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("EveOcean", 8.0, 1, 7));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("Gilly", 8.0, 1, 7));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("Ike", 8.0, 1, 7));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("Laythe", 8.0, 1, 10));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("LaytheOcean", 8.0, 1, 7));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("Moho", 8.0, 1, 9));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("Tylo", 8.0, 1, 9));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("Vall", 8.0, 1, 9));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("Dres", 8.0, 1, 9));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("Pol", 8.0, 1, 9));
-            pqspreset3.spherePresets.Add(new PQSCache.PQSSpherePreset("Eeloo", 8.0, 1, 9));
-            RuntimeUtility.RuntimeUtility.pqsLow = PQSCache.PresetList.presets[0];
-            RuntimeUtility.RuntimeUtility.pqsDefault = PQSCache.PresetList.presets[1];
-            RuntimeUtility.RuntimeUtility.pqsHigh = PQSCache.PresetList.presets[2];
-            PQSCache.PresetList.presets[0] = pqspreset;
-            PQSCache.PresetList.presets[1] = pqspreset2;
-            PQSCache.PresetList.presets[2] = pqspreset3;
-            return true;
-        }
-        static void Postfix(GameSettings __instance)
-        {
-            PQSCache.PresetList.presets[0] = RuntimeUtility.RuntimeUtility.pqsLow;
-            PQSCache.PresetList.presets[1] = RuntimeUtility.RuntimeUtility.pqsDefault;
-            PQSCache.PresetList.presets[2] = RuntimeUtility.RuntimeUtility.pqsHigh;
-        }
-    }
-    [HarmonyPatch(typeof(ModuleAsteroidInfo), "OnStart")]
-    public static class ModuleAsteroidInfo_OnStart
-    {
-        static bool Prefix(ModuleAsteroidInfo __instance, PartModule.StartState state)
-        {
-            __instance.baseMod = __instance.part.Modules.GetModule<ModuleAsteroid>(0);
-            if (__instance.baseMod != null)
-            {
-                __instance.baseMod.OnStart(state);
-                if (__instance.currentMassVal <= 1E-09)
-                {
-                    __instance.currentMassVal = (double)__instance.part.mass * RuntimeUtility.RuntimeUtility.KopernicusConfig.ApplyDensityMultToMinorObjects;
-                }
-                if (__instance.massThresholdVal <= 1E-09)
-                {
-                    __instance.SetupAsteroidResources();
-                }
-                __instance.part.force_activate();
-                __instance.baseMod.SetAsteroidMass((float)__instance.currentMassVal);
-                __instance.part.mass = (float)__instance.currentMassVal;
-            }
-            return false;
-        }
-    }
-
-    [HarmonyPatch(typeof(ModuleCometInfo), "OnStart")]
-    public static class ModuleCometInfo_OnStart
-    {
-        static bool Prefix(ModuleCometInfo __instance, PartModule.StartState state)
-        {
-            __instance.baseMod = __instance.part.Modules.GetModule<ModuleComet>(0);
-            if (__instance.baseMod != null)
-            {
-                __instance.baseMod.OnStart(state);
-                if (__instance.currentMassVal <= 1E-09)
-                {
-                    __instance.currentMassVal = (double)__instance.part.mass * RuntimeUtility.RuntimeUtility.KopernicusConfig.ApplyDensityMultToMinorObjects;
-                }
-                if (__instance.massThresholdVal <= 1E-09)
-                {
-                    __instance.SetupCometResources();
-                }
-                __instance.part.force_activate();
-                __instance.baseMod.SetCometMass((float)__instance.currentMassVal);
-                __instance.part.mass = (float)__instance.currentMassVal;
-            }
-            return false;
-        }
-    }
-
-    [HarmonyPatch(typeof(SentinelUtilities))]
-    [HarmonyPatch("FindInnerAndOuterBodies")]
-    [HarmonyPatch(new Type[] { typeof(double), typeof(CelestialBody), typeof(CelestialBody) }, new ArgumentType[] { ArgumentType.Normal, ArgumentType.Out, ArgumentType.Out })]
-    public static class SentinelUtilities_FindInnerAndOuterBodies
-    {
-        static bool Prefix(ref bool __result, double SMA, out CelestialBody innerBody, out CelestialBody outerBody)
-        {
-            Dictionary<double, CelestialBody> dictionary = new Dictionary<double, CelestialBody>();
-            for (int i = 0; i < FlightGlobals.Bodies.Count; i++)
-            {
-                CelestialBody celestialBody = FlightGlobals.Bodies[i];
-                if (celestialBody == Planetarium.fetch.Sun)
-                {
-                    dictionary.TryAdd(0.0, celestialBody);
-                }
-                else if (celestialBody.referenceBody == Planetarium.fetch.Sun)
-                {
-                    dictionary.TryAdd(celestialBody.orbit.semiMajorAxis, celestialBody);
-                }
-            }
-            List<double> list = dictionary.Keys.ToList<double>();
-            list.Sort();
-            for (int j = 0; j < list.Count; j++)
-            {
-                if (list[j] <= SMA)
-                {
-                    if (list[j + 1] > SMA)
-                    {
-                        innerBody = dictionary[list[j]];
-                        outerBody = dictionary[list[j + 1]];
-                        __result = true;
-                        return false;
-                    }
-                }
-            }
-            innerBody = dictionary[list[0]];
-            outerBody = dictionary[list[list.Count - 1]];
-            __result = false;
-            return false;
-        }
-    }
-
-    [HarmonyPatch(typeof(PQS), "UpdateVisual")]
-    internal static class PQS_UpdateVisual
-    {
-        static IEnumerable<CodeInstruction> Transpiler(IEnumerable<CodeInstruction> instructions)
-        {
-            var oldMethod = SymbolExtensions.GetMethodInfo<PQS>(pqs => pqs.GetSurfaceHeight(Vector3d.zero));
-            var newMethod = SymbolExtensions.GetMethodInfo<PQS>(pqs => GetSurfaceHeightIfNecessary(pqs, default));
-
-            var matcher = new CodeMatcher(instructions);
-            matcher
-                .MatchStartForward(new CodeMatch(OpCodes.Call, oldMethod))
-                .ThrowIfInvalid("Could not find call to PQS::GetSurfaceHeight")
-                .SetInstruction(new CodeInstruction(OpCodes.Call, newMethod));
-
-            return matcher.Instructions();
-        }
-
-        // PQS.Start calls GetSurfaceHeight for every body in the system, this includes ones that
-        // have no conceivable chance of being near the camera. This ends up being incredibly
-        // slow because it is necessary to load the heightmaps for all the planets in the system.
-        //
-        // This is a modified version of GetSurfaceHeight that skips returning the height if we're
-        // still in the loading screen and the planet is far enough away that it wouldn't be
-        // subdivided anyway.
-        //
-        // UpdateVisual gets called every frame during regular gameplay anyways, so this has no
-        // effect on anything outside of the main menu scene.
-        static double GetSurfaceHeightIfNecessary(PQS pqs, Vector3d radialVector)
-        {
-            if (HighLogic.LoadedScene != GameScenes.LOADING)
-                return pqs.GetSurfaceHeight(radialVector);
-
-            if (pqs.targetAltitude <= pqs.maxDetailDistance * pqs.radius)
-                return pqs.GetSurfaceHeight(radialVector);
-
-            return 0.0;
-        }
-    }
 }
